merge v17 - Remove translations using po header
Since an empty Comment in the metadata.desktop is treated as msgid "",
it conflicts with the po header. So when we call msgfmt to generate a
new desktop file, it uses the entire po header for the Comment[fr]
translation.

There doesn't seem to be a good way of telling xgettext to ignore
these empty messages, so we'll just delete the generated lines. It
will still complain that there's an empty msgid though.