Fix ConfigAdvanced duplicate "Default" keys.
Since KF5 5.78 (released 2021-01-02) (Debian 11 / Ubuntu 21.04),
ConfigPropertyMap loads the default values as
plasmoid.configuration.____Default

In recent versions of Qt, XHR on local files requires
QML_XHR_ALLOW_FILE_READ=1 which makes it useless to users for
debugging.

Since modelConfigType will not soon not work, we check if the number
is an integer in order to decide if we show decimals.