Support generating gettext compendiums based on installed kf5 messages
This used to work well with the OpenSUSE kf5 locale directory, but it
seems they moved most messages out of there in Plasma 6. I may need to
generate a list of KDE translation domains so that we only grab KDE
licensed translations.

Not sure why a bunch of compendium languages are reaching a memory
limit when I have 32Gb of RAM...

Also map 'kpac tr' and 'kpac translate' to 'kpac i18n'.